mogwai/scripts/requirements.txt
# Crystal Voice Server Dependencies
# Install with: pip install -r requirements.txt

# ── NVIDIA NeMo (primary — local GPU speech inference) ──────────
# Install PyTorch with CUDA first:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
# Then install NeMo:
nemo_toolkit[asr]>=2.0.0
torch>=2.1.0
huggingface_hub>=0.24

# Magpie TTS multilingual G2P dependencies
nltk>=3.8
jieba>=0.42
pypinyin>=0.50
pypinyin-dict>=0.8
pyopenjtalk>=0.3
janome>=0.5

# NVIDIA STT worker (WebSocket server)
websockets>=12.0
aiohttp>=3.9.0

# ── Whisper STT (fallback) ──────────────────────────────────────
faster-whisper>=1.0.0
ctranslate2>=4.0.0

# ── Kokoro TTS (fallback) ──────────────────────────────────────
kokoro>=0.9.2
# pyttsx3>=2.90  # Last-resort fallback

# ── Server ──────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# ── Audio processing ────────────────────────────────────────────
soundfile>=0.12.0
numpy>=1.24.0
